[Ref] DMO: Small changes.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@8005 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/plugins/dmo/Compressor.cpp
@@ -226,9 +226,9 @@ void Compressor::RecalculateCompressorParams()
 	m_gain = std::pow(10.0f, GainInDecibel() / 20.0f);
 	m_attack = std::pow(10.0f, -1.0f / (AttackTime() * sampleRate));
 	m_release = std::pow(10.0f, -1.0f / (ReleaseTime() * sampleRate));
-	const double _2e31 = double(1u << 31);
-	const double _2e26 = double(1u << 26);
-	m_threshold = static_cast<float>(std::min((_2e31 - 1.0), (std::log(std::pow(10.0f, ThresholdInDecibel() / 20.0f) * _2e31) * _2e26) / M_LN2 + _2e26) * (1.0 / _2e31));
+	const float _2e31 = float(1u << 31);
+	const float _2e26 = float(1u << 26);
+	m_threshold = std::min((_2e31 - 1.0f), (std::log(std::pow(10.0f, ThresholdInDecibel() / 20.0f) * _2e31) * _2e26) / static_cast<float>(M_LN2) + _2e26) * (1.0f / _2e31);
 	m_ratio = 1.0f - (1.0f / CompressorRatio());
 	m_predelay = static_cast<int32>((PreDelay() * sampleRate) + 2.0f);
 }
